+++ backend/parameters.py
@@ -42,7 +42,7 @@ class Parameters(CleepModule):
     """
 
     MODULE_AUTHOR = "Cleep"
-    MODULE_VERSION = "2.2.0"
+    MODULE_VERSION = "2.3.0"
     MODULE_CATEGORY = "APPLICATION"
     MODULE_DEPS = []
     MODULE_DESCRIPTION = "Configure generic parameters of your device"
@@ -620,7 +620,7 @@ class Parameters(CleepModule):
         Return non working days of current year
 
         Args:
-            date (int, optional): get non working day for specified year. If not specified use current year. Defaults to None.
+            year (int, optional): get non working day for specified year. If not specified use current year. Defaults to None.
 
         Returns:
             list: list of non working days of the year. List can be empty if error occured::
@@ -674,7 +674,6 @@ class Parameters(CleepModule):
 
         Args:
             day (str): day to check (must be iso format XXXX-MM-DD)
-            test (bool): new param
 
         Returns:
             bool: True if specified day is a non working day, False otherwise
@@ -726,7 +725,6 @@ class Parameters(CleepModule):
         Args:
             account (bool): account name
             password (str): account password (will be encrypted)
-            toto (bool, optional): new param. Defaults to False.
 
         Raises:
             CommandError: if adding account failed