+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [UNRELEASED]
+### Added
+- Handle secured access (aka auth)
+
 ## [2.2.0] - 2023-03-10
 ### Changed
 - Clean code